pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "journey2rich"
version = "0.1.0"
description = "Fundamental-driven trading agent for US equities and options (IBKR paper)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "xuzm15" }]

dependencies = [
  "pandas>=2.0",
  "numpy>=1.26",
  "ib_insync>=0.9.86",
  "yfinance>=0.2.40",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "typer>=0.12",
  "feedparser>=6.0",
  "apscheduler>=3.10",
]

[project.scripts]
journey2rich = "journey2rich.engine.runner:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
